fix(burstdownmix): Make offset a double before decimating

+++ lib/burst_downmix_impl.cc
@@ -383,7 +383,7 @@ namespace gr {
 
     int
     burst_downmix_impl::process_next_frame(float sample_rate, float center_frequency,
-            uint64_t offset, uint64_t id, size_t burst_size, int start)
+            double offset, uint64_t id, size_t burst_size, int start)
     {
       /*
        * Use the center frequency to make some assumptions about the burst.
@@ -618,7 +618,7 @@ namespace gr {
       float center_frequency = pmt::to_float(pmt::dict_ref(meta, pmt::mp("center_frequency"), pmt::PMT_NIL));
       float sample_rate = pmt::to_float(pmt::dict_ref(meta, pmt::mp("sample_rate"), pmt::PMT_NIL));
       uint64_t id = pmt::to_uint64(pmt::dict_ref(meta, pmt::mp("id"), pmt::PMT_NIL));
-      uint64_t offset = pmt::to_uint64(pmt::dict_ref(meta, pmt::mp("offset"), pmt::PMT_NIL));
+      double offset = pmt::to_uint64(pmt::dict_ref(meta, pmt::mp("offset"), pmt::PMT_NIL));
 
       if(id == d_debug_id) {
         d_debug = true;
@@ -628,7 +628,7 @@ namespace gr {
         printf("---------------> id:%" PRIu64 " len:%zu\n", id, burst_size);
         float absolute_frequency = center_frequency + relative_frequency * sample_rate;
         printf("relative_frequency=%f, absolute_frequency=%f\n", relative_frequency, absolute_frequency);
-        printf("offset=%" PRIu64 "\n", offset);
+        printf("offset=%f\n", offset);
         printf("sample_rate=%f\n", sample_rate);
       }
 
@@ -674,6 +674,7 @@ namespace gr {
       burst_size = burst_size / decimation;
 #else
       burst_size = (burst_size - d_input_fir.ntaps() + 1) / decimation;
+      //offset += d_input_fir.ntaps()/2;
 #endif
 
       d_input_fir.filterNdec(d_frame, d_tmp_a, burst_size, decimation);
